backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.3.4
    # via
    #   -r requirements.in
    #   aiohttp
aiohttp==3.10.1
    # via
    #   -r requirements.in
    #   django-nextjs
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
amqp==5.2.0
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   channels
    #   django
    #   django-cors-headers
attrs==24.1.0
    # via
    #   -r requirements.in
    #   aiohttp
billiard==4.2.0
    # via
    #   -r requirements.in
    #   celery
celery==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2024.7.4
    # via
    #   -r requirements.in
    #   requests
channels==4.1.0
    # via
    #   -r requirements.in
    #   django-nextjs
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements.in
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements.in
    #   celery
colorama==0.4.6
    # via
    #   -r requirements.in
    #   click
    #   tqdm
dataclasses==0.6
    # via -r requirements.in
django==5.0.7
    # via
    #   -r requirements.in
    #   channels
    #   django-celery-results
    #   django-cors-headers
    #   django-nextjs
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-results==2.5.1
    # via -r requirements.in
django-cors-headers==4.4.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-nextjs==3.2.0
    # via -r requirements.in
django-prometheus==2.3.1
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
dnspython==2.6.1
    # via -r requirements.in
filelock==3.15.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.1
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   -r requirements.in
    #   uvicorn
huggingface-hub==0.24.5
    # via
    #   -r requirements.in
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   -r requirements.in
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   torch
kombu==5.3.7
    # via
    #   -r requirements.in
    #   celery
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
multidict==6.0.5
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
networkx==3.3
    # via
    #   -r requirements.in
    #   torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   torchvision
    #   transformers
packaging==24.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
pillow==10.4.0
    # via
    #   -r requirements.in
    #   torchvision
prometheus-client==0.20.0
    # via
    #   -r requirements.in
    #   django-prometheus
prompt-toolkit==3.0.47
    # via
    #   -r requirements.in
    #   click-repl
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
redis==5.0.8
    # via
    #   -r requirements.in
    #   django-redis
regex==2024.7.24
    # via
    #   -r requirements.in
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
safetensors==0.4.4
    # via
    #   -r requirements.in
    #   transformers
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
sympy==1.13.1
    # via
    #   -r requirements.in
    #   torch
tokenizers==0.19.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.4.0
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
torchaudio==2.4.0
    # via -r requirements.in
torchvision==0.19.0
    # via -r requirements.in
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
transformers==4.43.4
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
tzdata==2024.1
    # via
    #   -r requirements.in
    #   celery
    #   django
urllib3==2.2.2
    # via
    #   -r requirements.in
    #   requests
uv==0.2.33
    # via -r requirements.in
uvicorn==0.30.5
    # via -r requirements.in
vine==5.1.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
websockets==12.0
    # via
    #   -r requirements.in
    #   django-nextjs
yarl==1.9.4
    # via
    #   -r requirements.in
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
